Add complex number and phase readouts to AudioAnalyzeFFT1024
The complex FFT used by the AnalyzeFFT1024 module yields complex numbers
both the amplitude and the phase of each frequency bin can be extracted
from. However, to this date the module only allows access to the
precomputed amplitudes. This commit adds the necessary API functions as
well as background code in order to make the following values of each
frequency bin available:
* The raw complex numbers
* Their real and imaginary parts
* Their unscaled magnitudes (previously stored in output[])
* The scaled amplitudes
* The scaled phases (-pi to pi)
Sadly, the old API functions and fields do longer fit into the new
naming scheme and thus have been marked as deprecated. Of course, they
do still work for the sake of backwards compatibility, even though they
are no longer documented. Moreover, this commit adjusts the examples and
GUI descriptions so that they reflect the changes as well.
Note that a similar extension of the AnalyzeFFT256 module might be
possible, but doesn't make much sense since the results would arrive in
such a quick succession that it might be impossible to process them
properly, both in the internal code as well as in the user's code.