Implement non-stalling option
Calling queue.setBehaviour(AudioPlayQueue::NON_STALLING) gives a non-zero result if ::play() did not complete; the call must be re-tried later.

In this mode ::getBuffer() may return NULL, as documented, and ::playBuffer() will return non-zero if it failed; as for ::play(), when this happens the operation must be re-tried later.