PTS8211 - attach interrupt/DMA.eanble before Transmit enable
There appears like there is a timing window where we were enabling the IS2 Transmit before we enabled the DMA and also had not set the interrupt vector.

So probably depending on what all is happening, we were running into cases where t5he QUEUE underrun was detected probably because of this.

Changed it for the two PTS8211 classes, reran @mjs513 sektch on post 93 in pjrc forum thread on Dual Serial, RAWHID, Serial+Midi16+Audio and a few others and they all appear to now make noise