Add dougcl's 64MB memory board; fix static initialization order fiasco
Only briefly tested with a partially populated 64MB memory board (see https://forum.pjrc.com/threads/29276-Limits-of-delay-effect-in-audio-library?p=320254&viewfull=1#post320254)

There was an issue with the undefined initialization order of the delay object(s) compared to the supporting SPI object. This wasn't apparent during previous testing, and may be a result of updates to the C++ compiler. We now "pre-initialise" the object, and only finish configuring SPI memory when the first call to delay() is made. Search for "C++ static initialization order fiasco" for more detail than you want.

Note that for huge amounts of SPI RAM the initialisation time can be quite significant, as the whole block is zeroed out at the start to ensure silence is output, before the first real samples are input to the memory.